Rename file according to vote in tools channel.
